Use a single SettingsProvider for public and private settings
