+++ client/admin/AdministrationRouter.js
@@ -1,17 +1,17 @@
 import React, { lazy, useMemo, Suspense } from 'react';
 
+import SettingsProvider from '../providers/SettingsProvider';
 import AdministrationLayout from './AdministrationLayout';
-import PrivilegedSettingsProvider from './PrivilegedSettingsProvider';
 import PageSkeleton from './PageSkeleton';
 
 function AdministrationRouter({ lazyRouteComponent, ...props }) {
 	const LazyRouteComponent = useMemo(() => lazy(lazyRouteComponent), [lazyRouteComponent]);
 	return <AdministrationLayout>
-		<PrivilegedSettingsProvider>
+		<SettingsProvider privileged>
 			<Suspense fallback={<PageSkeleton />}>
 				<LazyRouteComponent {...props} />
 			</Suspense>
-		</PrivilegedSettingsProvider>
+		</SettingsProvider>
 	</AdministrationLayout>;
 }
